Moved AddOnsDB to separate files
