+++ attacks.h
@@ -5,9 +5,8 @@ File name:  attacks.h
 Class:      CS202
 Term:	    Fall 2020
 
-This is the header file for the classes Attacks, Items, and AddOnsDb.
+This is the header file for the classes Attacks and Items.
 Items class is derived from Attacks class simply for dynamic binding in AddOnsDb.
-AddOnsDb require two .txt files in the specs folder to load attacks and items.
 */
 
 #ifndef ATTACKS_H
@@ -75,43 +74,6 @@ public:
 
 	void display(ostream &) const; // display data members
 	int useItem() const;		   // return hp of item for usage
-};
-
-class AddOnsDb
-{
-private:
-	Attacks **table;
-	int size;
-
-	void init(Attacks **, Attacks **, int);				  // initialize table to 0, array recursion
-	void copy(Attacks *&, Attacks *);					  // copy all table, list recursion
-	void copy(Attacks **, Attacks **, Attacks **);		  // copy all table, array recursion
-	void destroy(Attacks *&);							  // destroy everything, list recursion
-	void destroy(Attacks **);							  // destroy everything, array recursion
-	int insert(Attacks *&, Attacks *);					  // insert item, list recursion
-	bool insert(Attacks **, Attacks *);					  // insert item, array recursion
-	void display(Attacks *);							  // display all, list recursion
-	void display(Attacks **, int = 1);					  // display all, array recursion
-	Attacks *retrieve(Attacks *, int &, bool &, int = 0); // retrieve an attack, list recursion
-	Attacks *retrieve(Attacks **, const string &);		  // retrieve an attack, array recursion
-	int incrSize();										  // increase size of table array
-
-public:
-	AddOnsDb();					// default constructor
-	AddOnsDb(const AddOnsDb &); // copy constructor
-	~AddOnsDb();				// destructor
-
-	bool insert(const Attacks &);								// insert attack, helper function
-	bool insert(const Items &);									// insert item, helper function
-	void display();												// display all, helper function
-	Attacks *retrieve(const string &, const string = "normal"); // retrieve an attack by type
-	Attacks *retrieve();										// retrieve an item
-
-	bool addAttacks(const char *); // populate database with file of attacks
-	bool addItems(const char *);   // popluate database with file of items
-
-	AddOnsDb &operator+=(const Attacks &); //+= operator overload to add an attack
-	AddOnsDb &operator+=(const Items &);   //+= operator overload to add an item
 };
 
 #endif
