+++ makefile
@@ -1,4 +1,4 @@
-a.out: main.o minalib.o attacks.o pokemon.o gym.o
+a.out: main.o minalib.o attacks.o addonsdb.o pokemon.o gym.o
 	g++ -g *.o
 
 main.o: main.cpp
@@ -9,6 +9,9 @@ gym.o: gym.h gym.cpp
 
 pokemon.o: pokemon.h pokemon.cpp
 	g++ -g -c pokemon.cpp
+
+addonsdb.o: addonsdb.h addonsdb.cpp
+	g++ -g -c addonsdb.cpp
 
 attacks.o: attacks.h attacks.cpp
 	g++ -g -c attacks.cpp
